+++ kotlintest-assertions/src/main/kotlin/io/kotlintest/matchers/numerics/int.kt
@@ -3,12 +3,12 @@ package io.kotlintest.matchers.numerics
 import io.kotlintest.Matcher
 import io.kotlintest.Result
 
-fun isEven() = object : Matcher<Int> {
+fun even() = object : Matcher<Int> {
   override fun test(value: Int): Result =
       Result(value % 2 == 0, "$value should be even", "$value should be odd")
 }
 
-fun isOdd() = object : Matcher<Int> {
+fun odd() = object : Matcher<Int> {
   override fun test(value: Int): Result =
       Result(value % 2 == 1, "$value should be odd", "$value should be even")
 }