+++ kotlintest-tests/src/test/kotlin/com/sksamuel/kotlintest/tests/matchers/numerics/IntMatchersTest.kt
@@ -7,9 +7,14 @@ import io.kotlintest.matchers.beLessThanOrEqualTo
 import io.kotlintest.matchers.between
 import io.kotlintest.matchers.lt
 import io.kotlintest.matchers.lte
+import io.kotlintest.matchers.numerics.beInRange
+import io.kotlintest.matchers.numerics.even
+import io.kotlintest.matchers.numerics.odd
 import io.kotlintest.should
 import io.kotlintest.specs.StringSpec
 import io.kotlintest.shouldBe
+import io.kotlintest.shouldNot
+import io.kotlintest.shouldNotBe
 import io.kotlintest.shouldThrow
 import io.kotlintest.tables.forAll
 import io.kotlintest.tables.forNone
@@ -19,6 +24,24 @@ import io.kotlintest.tables.table
 
 class IntMatchersTest : StringSpec() {
   init {
+
+    "isEven" {
+      4 shouldBe even()
+      3 shouldNotBe even()
+    }
+
+    "isOdd" {
+      3 shouldBe odd()
+      4 shouldNotBe odd()
+    }
+
+    "inRange" {
+      3 should beInRange(1..10)
+      3 should beInRange(3..10)
+      3 should beInRange(3..3)
+      4 shouldNot beInRange(3..3)
+      4 shouldNot beInRange(1..3)
+    }
 
     "beGreaterThan" {
       1 should beGreaterThan(0)